requirements.txt
matplotlib==3.9.2
numpy==1.23.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==10.4.0
torch==2.4.0+cu124
torchvision==0.19.0
tqdm==4.66.4
